requirements.txt
absl-py==2.3.1
accelerate==1.10.1
agilerl==2.3.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
deepspeed==0.17.6
dill==0.3.9
distlib==0.4.0
docker-pycreds==0.4.0
einops==0.8.1
Farama-Notifications==0.0.4
fastrand==1.8.0
filelock==3.20.0
flatten-dict==0.4.2
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
gymnasium==1.2.1
h5py==3.14.0
hemac @ file:///mnt/data/Documents/Project_M/hemac
hf-xet==1.1.10
hjson==3.1.0
huggingface-hub==0.35.3
hydra-core==1.3.2
identify==2.6.15
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
minari==0.5.3
ml_dtypes==0.5.3
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
peft==0.15.2
pettingzoo==1.25.0
pillow==11.3.0
platformdirs==4.5.0
pre-commit==3.8.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pygame==2.6.1
Pygments==2.19.2
pymunk==6.11.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==4.6.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.41.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
stable_baselines3==2.7.0
SuperSuit==3.10.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.8.3
termcolor==1.1.0
tinyscaler==1.2.8
tokenizers==0.22.1
torch==2.5.1
tqdm==4.67.1
transformers==4.57.0
triton==3.1.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ucimlrepo==0.0.3
urllib3==2.5.0
virtualenv==20.35.3
wandb==0.17.9
Werkzeug==3.1.3
zipp==3.23.0
